requirements.txt
# ========================================
# RAG Agent 知识库系统 - Python 依赖
# ========================================

# --- 向量数据库 ---
chromadb>=0.5.0

# --- OpenAI ---
openai>=1.30.0

# --- LangChain 生态（含 LangGraph Agent 编排）---
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0          # ReAct Agent 编排核心

# --- GitHub API ---
PyGithub>=2.3.0
gitpython>=3.1.40

# --- 文档处理 ---
tiktoken>=0.7.0           # Token计数
pypdf>=4.0.0              # PDF解析
python-docx>=1.1.0        # Word文档
beautifulsoup4>=4.12.0    # HTML解析
markdown>=3.6.0           # Markdown处理

# --- Web界面 ---
streamlit>=1.35.0

# --- 工具库 ---
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0              # 美化命令行输出
tenacity>=8.3.0           # 重试机制
httpx>=0.27.0
tqdm>=4.66.0              # 进度条

# --- 数据处理 ---
pandas>=2.2.0
numpy>=1.26.0
